moved Vue Components to /components folder
